Remove chain2 merge into chain
